Feat: utils header created.

+++ include/dijkstra.hpp
@@ -3,8 +3,9 @@
 
 #include "graph.hpp"
 #include "tuple.hpp"
+#include "utils.hpp"
 
 double dijkstra(Graph<Tuple<double, double>, double> &trail, Graph<Tuple<double, double>, double> &portals,
-long long source, long long dest, long long glades, long long portals_allowed);
+ll source, ll dest, ll glades, ll portals_allowed);
 
 #endif
