+++ include/heap.hpp
@@ -2,12 +2,13 @@
 #define HEAP_HPP
 
 #include "tuple.hpp"
+#include "utils.hpp"
 
 template <typename Key, typename Value, class Comparator, class Hash, typename HashParam>
 class Heap{
 
     public:
-        Heap(long long maxsize, HashParam hashparam);
+        Heap(ll maxsize, HashParam hashparam);
         ~Heap();
 
         void insert(Key key, Value value);
@@ -18,20 +19,20 @@ class Heap{
         bool empty();
 
     private:
-        long long getFather(long long pos);
-        long long getLeftChild(long long pos);
-        long long getRightChild(long long pos);
+        ll getFather(ll pos);
+        ll getLeftChild(ll pos);
+        ll getRightChild(ll pos);
 
         Comparator comp;
         Hash hash;
 
-        long long size;
-        long long maxSize;
+        ll size;
+        ll maxSize;
         Tuple<Key, Value> *data;
-        long long *indexes;
+        ll *indexes;
 
-        void heapifyDown(long long pos);
-        void heapifyUp(long long pos);
+        void heapifyDown(ll pos);
+        void heapifyUp(ll pos);
 };
