+++ src/hash.cpp
@@ -1,7 +1,7 @@
 #include "hash.hpp"
 
-Hash_Lost_Woods::Hash_Lost_Woods(Tuple<long long, long long> vertice) : glades(vertice.first), portals(vertice.second) {}
+Hash_Lost_Woods::Hash_Lost_Woods(Tuple<ll, ll> vertice) : glades(vertice.first), portals(vertice.second) {}
 
-long long Hash_Lost_Woods::operator() (const Tuple<long long, long long>& key) const {
+ll Hash_Lost_Woods::operator() (const Tuple<ll, ll>& key) const {
     return (glades * key.first + key.second);
 }
