+++ src/main.cpp
@@ -1,12 +1,13 @@
 #include <iostream>
 #include "heap.hpp"
 #include "graph.hpp"
+#include "utils.hpp"
 #include "euclidean_distance.hpp"
 #include "dijkstra.hpp"
 
 int main() {
 
-    long long n, m, k; 
+    ll n, m, k; 
     std::cin >> n >> m >> k;
 
     Graph<Tuple<double, double>, double> trail(n);
@@ -24,7 +25,7 @@ int main() {
 
     for (size_t i = 0; i < m; i++)
     {
-        Tuple<long long, long long> vertices;
+        Tuple<ll, ll> vertices;
         std::cin >> vertices.first >> vertices.second;
 
         trail.insertEdge(vertices.first, vertices.second, euclidean_distance(trail.getVertice(vertices.first), trail.getVertice(vertices.second)));
@@ -32,13 +33,13 @@ int main() {
 
     for (size_t i = 0; i < k; i++)
     {
-        Tuple<long long, long long> vertices;
+        Tuple<ll, ll> vertices;
         std::cin >> vertices.first >> vertices.second;
 
         portals.insertEdge(vertices.first, vertices.second, 0);
     }   
 
-    double s; long long q;
+    double s; ll q;
     std::cin >> s >> q;
 
     std::cout << dijkstra(trail, portals, 0, n-1, n, q) << '\n';
