+++ src/tuple.cpp
@@ -1,4 +1,5 @@
 #include "tuple.hpp"
+#include "utils.hpp"
 
 template<typename T1,typename T2>
 Tuple<T1, T2>::Tuple() : first(T1()), second(T2()) {}
@@ -11,8 +12,8 @@ bool Tuple<T1, T2>::operator==(const Tuple<T1, T2>& other) const {
     return (this->first == other.first && this->second == other.second);
 }
 
-template class Tuple<long long, long long>;
-template class Tuple<long long, int>;
-template class Tuple<long long, double>;
+template class Tuple<ll, ll>;
+template class Tuple<ll, int>;
+template class Tuple<ll, double>;
 template class Tuple<double, double>;
-template class Tuple<Tuple<long long, long long>, double>;
+template class Tuple<Tuple<ll, ll>, double>;
